fix linter errors and convert TagsAddButton and ConfirmButton to typescript
