+++ ui/src/types/kapacitor.ts
@@ -195,3 +195,7 @@ export interface Input {
 export interface Checkbox {
   checked: boolean
 }
+
+export interface DropdownItem {
+  text: string
+}
